+++ src/lib/isURL.js
@@ -14,6 +14,8 @@ const default_url_options = {
   allow_protocol_relative_urls: false,
 };
 
+const wrapped_ipv6 = /^\[([^\]]+)\](?::([0-9]+))?$/;
+
 export default function isURL(url, options) {
   assertString(url);
   if (!url || url.length >= 2083 || /\s/.test(url)) {
@@ -23,7 +25,7 @@ export default function isURL(url, options) {
     return false;
   }
   options = merge(options, default_url_options);
-  let protocol, auth, host, hostname, port, port_str, split;
+  let protocol, auth, host, hostname, port, port_str, split, ipv6;
 
   split = url.split('#');
   url = split.shift();
@@ -54,16 +56,29 @@ export default function isURL(url, options) {
     }
   }
   hostname = split.join('@');
-  split = hostname.split(':');
-  host = split.shift();
-  if (split.length) {
-    port_str = split.join(':');
+
+  port_str = ipv6 = null;
+  const ipv6_match = hostname.match(wrapped_ipv6);
+  if (ipv6_match) {
+    host = '';
+    ipv6 = ipv6_match[1];
+    port_str = ipv6_match[2] || null;
+  } else {
+    split = hostname.split(':');
+    host = split.shift();
+    if (split.length) {
+      port_str = split.join(':');
+    }
+  }
+
+  if (port_str !== null) {
     port = parseInt(port_str, 10);
     if (!/^[0-9]+$/.test(port_str) || port <= 0 || port > 65535) {
       return false;
     }
   }
-  if (!isIP(host) && !isFQDN(host, options) &&
+
+  if (!isIP(host) && !isFQDN(host, options) && (!ipv6 || !isIP(ipv6, 6)) &&
           host !== 'localhost') {
     return false;
   }