+++ validator.js
@@ -295,6 +295,7 @@
         protocols: ['http', 'https', 'ftp'],
         require_tld: true,
         require_protocol: false,
+        require_host: true,
         require_valid_protocol: true,
         allow_underscores: false,
         allow_trailing_dot: false,
@@ -342,6 +343,11 @@
 
         split = url.split('/');
         url = split.shift();
+
+        if (url === '' && !options.require_host) {
+          return true;
+        }
+
         split = url.split('@');
         if (split.length > 1) {
           auth = split.shift();