+++ validator.js
@@ -1142,7 +1142,7 @@
         return parts.join('@');
       }
 
-      var version = '5.6.0';
+      var version = '5.7.0';
 
       var validator = {
         version: version,