+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added support for IPv6 in `isURL()`
   ([#564](https://github.com/chriso/validator.js/issues/564))
+- Added support for urls without a host (e.g. `file:///foo.txt`) in `isURL()`
+  ([#563](https://github.com/chriso/validator.js/issues/563))
 - Added support for MasterCard 2-Series BIN
   ([#576](https://github.com/chriso/validator.js/pull/576))
 - New locales